pre-deployment commit

+++ src/pages/signUpUpdate.jsx
@@ -5,6 +5,7 @@ import { useUserContext } from '../context/userContext'
 import { addUser } from '../connections/firebase'
 import { signUpUser } from '../connections/firebase'
 import { Link } from 'react-router-dom';
+import { CircularProgress } from '@mui/material'
 
 
 const SignUpUpdate = () => {
@@ -15,6 +16,7 @@ const SignUpUpdate = () => {
   const [confirmPassword, setConfirmPassword] = useState('');
   const { dispatch } = useAuth()
   const { dispatch: userDispatch } = useUserContext()
+  const [isLoading, setIsLoading] = useState(false)
 
   const handleSubmit = (e) => {
     e.preventDefault();
@@ -23,7 +25,7 @@ const SignUpUpdate = () => {
   };
 
   return (
-    <div className= 'signupPage'>
+    <div className='signupPage'>
       <div className="signup-container">
         <h1>Sign up</h1>
         <p>Create your account</p>
@@ -61,7 +63,7 @@ const SignUpUpdate = () => {
           <div className="input-group">
             <i className='icon'>&#x1F512;</i>
             <input
-              type="text"
+              type="password"
               placeholder="Password"
               value={password}
               onChange={(e) => setPassword(e.target.value)}
@@ -71,8 +73,8 @@ const SignUpUpdate = () => {
           <div className="input-group">
             <i>&#x1F512;</i>
             <input
-              type="text"
               placeholder="Confirm Password"
+              type="password"
               value={confirmPassword}
               onChange={(e) => setConfirmPassword(e.target.value)}
               required
@@ -83,29 +85,36 @@ const SignUpUpdate = () => {
             className="signup-button"
             onClick={async function (e) {
               e.preventDefault()
-              console.log({ username, email, password, confirmPassword, phoneNumber })
+              if (password !== confirmPassword) {
+                alert("Passwords do not match")
+                return
+              }
+              console.log({ username, email, password, phoneNumber })
+              setIsLoading(true)
               try {
                 const signedUpUser = await signUpUser(email, password)
                 localStorage.setItem("emsToken", signedUpUser.accessToken)
                 console.log("user in the signup", signedUpUser)
-                const docsnap = await addUser(signedUpUser.uid, username, email, password, confirmPassword, phoneNumber)
+                const docsnap = await addUser(signedUpUser.uid, username, email, password, phoneNumber)
                 if (docsnap !== undefined) {
                   dispatch({ type: 'LOGIN' })
                   userDispatch({ type: 'SET_USER', payload: docsnap })
 
                 }
-
+                setIsLoading(false)
               } catch (error) {
+                setIsLoading(false)
                 console.log(error)
+                alert("An error occured")
               }
             }}
           >
-            Sign up
+              {isLoading ? <CircularProgress sx={{color: "white"}}/> : " Sign up"}
           </button>
         </form>
         <div className="or">Or</div>
         <button className="google-login">
-           Login with Google
+          Login with Google
         </button>
         <div className="login-link">
           Already have an account? <span> <Link style={{ color: '#d1410c' }} to='/loginUpdate'>Login</Link> </span>
